CAUTION WIP: Move toolhead above bed before probing
CAUTION: Because this is on a Positron V3.2 batch 2, we're taking advantage
of the fact we have an endstop here to figure out if this idea actually
will help.

Basically, we want to move close to the bed before actually probing so that
we can potentially use a much lower threshold and lower speed for probes.